LayoutTests/storage/domstorage/localstorage/blocked-file-access.html

<html>
<head>
<script>
if (window.testRunner) {
    testRunner.setAllowUniversalAccessFromFileURLs(false);
    testRunner.dumpAsText();
    testRunner.waitUntilDone();
}
</script>
</head>
<body>
<iframe src="resources/blocked-example.html"></iframe>
<div id="results"></div>
Test that we cannot access localStorage from a file URL if unversal access is turned off.
</body>
</html>
